ofz#820 oom terminate
comments claim return is len of data, but negative numbers for compressed data errors
are returned. Make things return the amount of successful data read and turn into
unsigned size_t to match ofz#727

Reviewed-on: https://gerrit.libreoffice.org/35079
Tested-by: Jenkins <ci@libreoffice.org>
(cherry picked from commit 5cff7d887f249ff85850fe441e0f564bdcde1492)

+++ hwpfilter/source/drawing.h
@@ -107,14 +107,14 @@ static bool SkipPrivateBlock(int type)
     {
         if (!hmem->read4b(n))
             return false;
-        if (hmem->state() || hmem->skipBlock(n) != n)
+        if (hmem->state() || hmem->skipBlock(n) != static_cast<size_t>(n))
             return false;
     }
     if (!hmem->read4b(n))
         return false;
     if (hmem->state())
         return false;
-    return hmem->skipBlock(n) == n;
+    return hmem->skipBlock(n) == static_cast<size_t>(n);
 }
 
 static int SizeExpected;
